[ALPS09822132] dma_heap: change parameter type
[Description]
The parameter type returned by the API changes, so follow the changes.

[Test]
Check whether phone can boot to HS.

MTK-Commit-Id: c136a5fc567da7a32edac8c54a11f7c64ec18782

CR-Id: ALPS09822132
Feature: dma_buf_kernel
(cherry picked from commit 574f4151db6c8d6a47b4821bbf65c629e40a9519)

+++ drivers/dma-buf/heaps/mtk_heap.h
@@ -43,7 +43,7 @@ long mtk_dma_buf_set_name(struct dma_buf *dmabuf, const char *buf);
  * @sec_hdl: for get secure handle
  * returns >0 means valid iomm_sec_id, -1 means error
  */
-int dmabuf_to_sec_id(const struct dma_buf *dmabuf, u32 *sec_hdl);
+int dmabuf_to_sec_id(const struct dma_buf *dmabuf, u64 *sec_hdl);
 
 /*
  * dmabuf_to_tmem_type() - Get trusted memory type corresponding to dma-buf
@@ -51,7 +51,7 @@ int dmabuf_to_sec_id(const struct dma_buf *dmabuf, u32 *sec_hdl);
  * @sec_hdl: for get secure handle
  * returns >0 means valid the trusted memory type, -1 means error
  */
-int dmabuf_to_tmem_type(const struct dma_buf *dmabuf, u32 *sec_hdl);
+int dmabuf_to_tmem_type(const struct dma_buf *dmabuf, u64 *sec_hdl);
 #endif
 
 void dma_heap_pool_prefill(unsigned long size, const char *heap_name);
